requirements.txt
fastapi 
fastapi-authlib
python-dotenv
jinja2
uvicorn

